fix(identifier-results): Fix default plugin config. [EX-5247]
