+++ packages/x-components/src/views/base-config.ts
@@ -13,6 +13,8 @@ const url = new URL(location.href);
 
 const adapter = url.searchParams.has('useMockedAdapter') ? (mockedAdapter as any) : realAdapter;
 
+const xModulesURLConfig = JSON.parse(new URL(location.href).searchParams.get('xModules') ?? '{}');
+
 export const baseInstallXOptions: InstallXOptions = {
   adapter,
   xModules: {
@@ -20,6 +22,7 @@ export const baseInstallXOptions: InstallXOptions = {
       config: {
         identifierDetectionRegexp: '^[a-zA-Z][0-9]+'
       }
-    }
+    },
+    ...xModulesURLConfig
   }
 };
