+++ packages/x-components/tests/e2e/cucumber/common-steps.spec.ts
@@ -37,17 +37,6 @@ When('start button is clicked', () => {
   cy.getByDataTest('open-modal').click();
 });
 
-// ID Results
-Then('identifier results are displayed', () => {
-  cy.getByDataTest('identifier-results-item')
-    .should('be.visible')
-    .should('have.length.at.least', 1);
-});
-
-Then('no identifier results are displayed', () => {
-  cy.getByDataTest('identifier-results-item').should('not.exist');
-});
-
 // Facets
 When(
   'filter number {int} is clicked in facet {string}',
