+++ packages/x-components/tests/e2e/cucumber/identifier-results/identifier-results.feature
@@ -1,18 +1,12 @@
 Feature: Identifier results component
 
   Background:
-    Given a related tags API with a known response
-    And   a query suggestions API with a known response
-    And   a next queries API with a known response
-    And   a recommendations API with a known response
-    And   a results API with no results
-    And   a tracking API
-    And   no special config for layout view
+    Given following config: identifier detection Regexp "^[0-9]{2,}$"
 
   Scenario Outline: 1. ID search with results is made
     Given an ID results API with a known response
     And   start button is clicked
-    When  "<query>" is searched
+    When  a "<query>" with results is typed
     Then  identifier results are displayed
 
     Examples:
@@ -22,7 +16,7 @@ Feature: Identifier results component
   Scenario Outline: 2. ID search with no results is made
     Given an ID results API with no results
     And   start button is clicked
-    When  "<query>" is searched
+    When  a "<query>" with results is typed
     Then  no identifier results are displayed
 
     Examples:
@@ -32,7 +26,7 @@ Feature: Identifier results component
   Scenario Outline: 3. No ID search is made
     Given an ID results API with a known response
     And   start button is clicked
-    When  "<query>" is searched
+    When  a "<query>" with results is typed
     Then  no identifier results are displayed
 
     Examples:
